:recycle: refactor:Extract functions
