+++ lib/screens/transactions/transactions_list.dart
@@ -1,10 +1,13 @@
 import 'package:bytebank/components/centered_message.dart';
 import 'package:bytebank/components/progress.dart';
 import 'package:bytebank/models/transaction.dart';
-import 'package:bytebank/services/api.dart';
+import 'package:bytebank/services/routes/transactions_routes.dart';
 import 'package:flutter/material.dart';
 
 class TransactionsList extends StatelessWidget {
+
+  final TransactionRoutes _transactionRoute = TransactionRoutes();
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -12,7 +15,7 @@ class TransactionsList extends StatelessWidget {
         title: Text('Transactions'),
       ),
       body: FutureBuilder<List<Transaction>>(
-        future: findAll(),
+        future: _transactionRoute.findAll(),
         builder: (context, snapshot) {
           switch (snapshot.connectionState) {
             case ConnectionState.none:
