+++ lib/services/api.dart
@@ -1,30 +1,10 @@
-import 'dart:convert';
 
-import 'package:bytebank/models/contact.dart';
-import 'package:bytebank/models/transaction.dart';
 import 'package:http/http.dart';
 import 'package:http_interceptor/http_interceptor.dart';
 
-class LoggingInterceptor implements InterceptorContract {
-  @override
-  Future<RequestData> interceptRequest({RequestData data}) async {
-    print('Request');
-    print('url ${data.url}');
-    print('headers ${data.headers}');
-    print('body ${data.body}');
-    return data;
-  }
+import 'interceptors/logging_interceptor.dart';
 
-  @override
-  Future<ResponseData> interceptResponse({ResponseData data}) async {
-    print('Response');
-    print('Status code: ${data.statusCode}');
-    print('url ${data.url}');
-    print('headers ${data.headers}');
-    print('body ${data.body}');
-    return data;
-  }
-}
+
 
 final Client client = HttpClientWithInterceptor.build(
   interceptors: [LoggingInterceptor()],
@@ -32,51 +12,4 @@ final Client client = HttpClientWithInterceptor.build(
 
 const String baseUrl = '2f42a37ec63a.ngrok.io';
 
-Future<List<Transaction>> findAll() async {
-  final Response response = await client
-      .get(Uri.https(baseUrl, 'transactions'))
-      .timeout(Duration(seconds: 5));
 
-  final List<dynamic> decodedJson = jsonDecode(response.body);
-  final List<Transaction> transactions = [];
-  for (Map<String, dynamic> transactionJson in decodedJson) {
-    final Map<String, dynamic> contactJson = transactionJson['contact'];
-    final Transaction transaction = Transaction(
-      transactionJson['value'],
-      Contact(
-        0,
-        contactJson['name'],
-        contactJson['accountNumber'],
-      ),
-    );
-    transactions.add(transaction);
-  }
-  return transactions;
-}
-
-Future<Transaction> save(Transaction transaction) async {
-  final Map<String, dynamic> transactionMap = {
-    'value': transaction.value,
-    'contact': {
-      'name': transaction.contact.name,
-      'accountNumber': transaction.contact.account
-    }
-  };
-  final String transactionJson = jsonEncode(transactionMap);
-  final Response response = await client.post(
-      Uri.https(baseUrl, 'transactions'),
-      headers: {'Content-type': 'application/json', 'password': '1000'},
-      body: transactionJson);
-
-  Map<String, dynamic> json = jsonDecode(response.body);
-
-  final Map<String, dynamic> contactJson = json['contact'];
-  return Transaction(
-    json['value'],
-    Contact(
-      0,
-      contactJson['name'],
-      contactJson['accountNumber'],
-    ),
-  );
-}
