+++ lib/services/routes/transactions_routes.dart
@@ -1,0 +1,72 @@
+import 'dart:convert';
+
+import 'package:bytebank/models/contact.dart';
+import 'package:bytebank/models/transaction.dart';
+import 'package:http/http.dart';
+
+import '../api.dart';
+
+class TransactionRoutes{
+  Future<List<Transaction>> findAll() async {
+    final Response response = await client
+        .get(Uri.https(baseUrl, 'transactions'))
+        .timeout(Duration(seconds: 5));
+
+    List<Transaction> transactions = _toTransactions(response);
+    return transactions;
+  }
+
+  List<Transaction> _toTransactions(Response response) {
+    final List<dynamic> decodedJson = jsonDecode(response.body);
+    final List<Transaction> transactions = [];
+    for (Map<String, dynamic> transactionJson in decodedJson) {
+      final Map<String, dynamic> contactJson = transactionJson['contact'];
+      final Transaction transaction = Transaction(
+        transactionJson['value'],
+        Contact(
+          0,
+          contactJson['name'],
+          contactJson['accountNumber'],
+        ),
+      );
+      transactions.add(transaction);
+    }
+    return transactions;
+  }
+
+  Future<Transaction> save(Transaction transaction) async {
+    Map<String, dynamic> transactionMap = _toMap(transaction);
+    final String transactionJson = jsonEncode(transactionMap);
+    final Response response = await client.post(
+        Uri.https(baseUrl, 'transactions'),
+        headers: {'Content-type': 'application/json', 'password': '1000'},
+        body: transactionJson);
+
+    return _toTransaction(response);
+  }
+
+  Transaction _toTransaction(Response response) {
+    Map<String, dynamic> json = jsonDecode(response.body);
+
+    final Map<String, dynamic> contactJson = json['contact'];
+    return Transaction(
+      json['value'],
+      Contact(
+        0,
+        contactJson['name'],
+        contactJson['accountNumber'],
+      ),
+    );
+  }
+
+  Map<String, dynamic> _toMap(Transaction transaction) {
+    final Map<String, dynamic> transactionMap = {
+      'value': transaction.value,
+      'contact': {
+        'name': transaction.contact.name,
+        'accountNumber': transaction.contact.account
+      }
+    };
+    return transactionMap;
+  }
+}
